refactor(riscv): split loadstore tests by width
