+++ extensions/riscv/circuit/src/load_sign_extend/tests/mod.rs
@@ -27,7 +27,6 @@ use openvm_stark_backend::{
 };
 use openvm_stark_sdk::{p3_baby_bear::BabyBear, utils::create_seeded_rng};
 use rand::{rngs::StdRng, Rng};
-use test_case::test_case;
 #[cfg(feature = "cuda")]
 use {
     super::{
@@ -74,6 +73,10 @@ use crate::{
 const IMM_BITS: usize = 16;
 const MAX_INS_CAPACITY: usize = 128;
 type F = BabyBear;
+
+mod byte;
+mod halfword;
+mod word;
 
 type ByteHarness = TestChipHarness<
     F,
@@ -279,153 +282,6 @@ fn memory_config_for() -> MemoryConfig {
     mem_config.addr_spaces[PUBLIC_VALUES_AS as usize].num_cells = 1 << 29;
     mem_config
 }
-
-#[test_case(LOADB, 100)]
-fn rand_load_sign_extend_byte_test(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for());
-    let (mut harness, bitwise) = create_byte_harness(&mut tester);
-    for _ in 0..num_ops {
-        set_and_execute(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.arena,
-            &mut rng,
-            opcode,
-            None,
-            None,
-            None,
-        );
-    }
-    tester
-        .build()
-        .load(harness)
-        .load_periphery(bitwise)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[test_case(LOADH, 100)]
-fn rand_load_sign_extend_halfword_test(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for());
-    let mut harness = create_halfword_harness(&mut tester);
-    for _ in 0..num_ops {
-        set_and_execute(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.arena,
-            &mut rng,
-            opcode,
-            None,
-            None,
-            None,
-        );
-    }
-    tester
-        .build()
-        .load(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[test_case(LOADW, 100)]
-fn rand_load_sign_extend_word_test(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for());
-    let mut harness = create_word_harness(&mut tester);
-    for _ in 0..num_ops {
-        set_and_execute(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.arena,
-            &mut rng,
-            opcode,
-            None,
-            None,
-            None,
-        );
-    }
-    tester
-        .build()
-        .load(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[test]
-fn positive_loadb_shift7_test() {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for());
-    let (mut harness, bitwise) = create_byte_harness(&mut tester);
-    set_and_execute(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADB,
-        Some([7, 0, 0, 0, 0, 0, 0, 0]),
-        Some(0),
-        Some(0),
-    );
-    tester
-        .build()
-        .load(harness)
-        .load_periphery(bitwise)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[test]
-fn positive_loadh_shift6_test() {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for());
-    let mut harness = create_halfword_harness(&mut tester);
-    set_and_execute(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADH,
-        Some([6, 0, 0, 0, 0, 0, 0, 0]),
-        Some(0),
-        Some(0),
-    );
-    tester
-        .build()
-        .load(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[test]
-fn positive_loadw_shift4_test() {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for());
-    let mut harness = create_word_harness(&mut tester);
-    set_and_execute(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADW,
-        Some([4, 0, 0, 0, 0, 0, 0, 0]),
-        Some(0),
-        Some(0),
-    );
-    tester
-        .build()
-        .load(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
 fn assert_pranked_byte_fails(prank: impl Fn(&mut LoadSignExtendByteCoreCols<F>)) {
     let mut rng = create_seeded_rng();
     let mut tester = VmChipTestBuilder::from_config(memory_config_for());
@@ -533,149 +389,12 @@ fn negative_split_signed_load_tests() {
 //
 //  Ensure GPU tracegen is equivalent to CPU tracegen.
 // ////////////////////////////////////////////////////////////////////////////////////
-
-#[cfg(feature = "cuda")]
-type GpuByteHarness = GpuTestChipHarness<
-    F,
-    Rv64LoadSignExtendByteExecutor,
-    Rv64LoadSignExtendByteAir,
-    Rv64LoadSignExtendByteChipGpu,
-    Rv64LoadSignExtendByteChip<F>,
->;
-
-#[cfg(feature = "cuda")]
-type GpuHalfwordHarness = GpuTestChipHarness<
-    F,
-    Rv64LoadSignExtendHalfwordExecutor,
-    Rv64LoadSignExtendHalfwordAir,
-    Rv64LoadSignExtendHalfwordChipGpu,
-    Rv64LoadSignExtendHalfwordChip<F>,
->;
-
-#[cfg(feature = "cuda")]
-type GpuWordHarness = GpuTestChipHarness<
-    F,
-    Rv64LoadSignExtendWordExecutor,
-    Rv64LoadSignExtendWordAir,
-    Rv64LoadSignExtendWordChipGpu,
-    Rv64LoadSignExtendWordChip<F>,
->;
-
 #[cfg(feature = "cuda")]
 fn dummy_range_checker() -> Arc<VariableRangeCheckerChip> {
     Arc::new(VariableRangeCheckerChip::new(
         default_var_range_checker_bus(),
     ))
 }
-
-#[cfg(feature = "cuda")]
-fn create_cuda_byte_harness(tester: &GpuChipTestBuilder) -> GpuByteHarness {
-    let range_checker = dummy_range_checker();
-    let bitwise_chip = Arc::new(BitwiseOperationLookupChip::<RV64_BYTE_BITS>::new(
-        default_bitwise_lookup_bus(),
-    ));
-    let air = Rv64LoadSignExtendByteAir::new(
-        Rv64LoadStoreAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadSignExtendByteCoreAir::new(
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            bitwise_chip.bus(),
-            range_checker.bus(),
-        ),
-    );
-    let executor = Rv64LoadSignExtendByteExecutor::new(
-        Rv64LoadStoreAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let cpu_chip = Rv64LoadSignExtendByteChip::<F>::new(
-        LoadSignExtendByteFiller::new(
-            Rv64LoadStoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            bitwise_chip,
-            range_checker,
-        ),
-        tester.dummy_memory_helper(),
-    );
-    let gpu_chip = Rv64LoadSignExtendByteChipGpu::new(
-        tester.range_checker(),
-        tester.bitwise_op_lookup(),
-        tester.address_bits(),
-        tester.timestamp_max_bits(),
-    );
-
-    GpuTestChipHarness::with_capacity(executor, air, gpu_chip, cpu_chip, MAX_INS_CAPACITY)
-}
-
-#[cfg(feature = "cuda")]
-fn create_cuda_halfword_harness(tester: &GpuChipTestBuilder) -> GpuHalfwordHarness {
-    let range_checker = dummy_range_checker();
-    let air = Rv64LoadSignExtendHalfwordAir::new(
-        Rv64LoadStoreAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadSignExtendHalfwordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET, range_checker.bus()),
-    );
-    let executor = Rv64LoadSignExtendHalfwordExecutor::new(
-        Rv64LoadStoreAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let cpu_chip = Rv64LoadSignExtendHalfwordChip::<F>::new(
-        LoadSignExtendHalfwordFiller::new(
-            Rv64LoadStoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.dummy_memory_helper(),
-    );
-    let gpu_chip = Rv64LoadSignExtendHalfwordChipGpu::new(
-        tester.range_checker(),
-        tester.address_bits(),
-        tester.timestamp_max_bits(),
-    );
-
-    GpuTestChipHarness::with_capacity(executor, air, gpu_chip, cpu_chip, MAX_INS_CAPACITY)
-}
-
-#[cfg(feature = "cuda")]
-fn create_cuda_word_harness(tester: &GpuChipTestBuilder) -> GpuWordHarness {
-    let range_checker = dummy_range_checker();
-    let air = Rv64LoadSignExtendWordAir::new(
-        Rv64LoadStoreAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadSignExtendWordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET, range_checker.bus()),
-    );
-    let executor = Rv64LoadSignExtendWordExecutor::new(
-        Rv64LoadStoreAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let cpu_chip = Rv64LoadSignExtendWordChip::<F>::new(
-        LoadSignExtendWordFiller::new(
-            Rv64LoadStoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.dummy_memory_helper(),
-    );
-    let gpu_chip = Rv64LoadSignExtendWordChipGpu::new(
-        tester.range_checker(),
-        tester.address_bits(),
-        tester.timestamp_max_bits(),
-    );
-
-    GpuTestChipHarness::with_capacity(executor, air, gpu_chip, cpu_chip, MAX_INS_CAPACITY)
-}
-
 #[cfg(feature = "cuda")]
 fn transfer_load_sign_extend_records<G, C, A, E>(harness: &mut GpuTestChipHarness<F, E, A, G, C>) {
     type Record<'a> = (&'a mut Rv64LoadStoreAdapterRecord, &'a mut LoadStoreRecord);
@@ -686,86 +405,4 @@ fn transfer_load_sign_extend_records<G, C, A, E>(harness: &mut GpuTestChipHarnes
             &mut harness.matrix_arena,
             EmptyAdapterCoreLayout::<F, Rv64LoadStoreAdapterExecutor>::new(),
         );
-}
-
-#[cfg(feature = "cuda")]
-#[test_case(LOADB, 100)]
-fn test_cuda_rand_load_sign_extend_byte_tracegen(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester =
-        GpuChipTestBuilder::default().with_bitwise_op_lookup(default_bitwise_lookup_bus());
-    let mut harness = create_cuda_byte_harness(&tester);
-    for _ in 0..num_ops {
-        set_and_execute(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.dense_arena,
-            &mut rng,
-            opcode,
-            None,
-            None,
-            None,
-        );
-    }
-    transfer_load_sign_extend_records(&mut harness);
-    tester
-        .build()
-        .load_gpu_harness(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[cfg(feature = "cuda")]
-#[test_case(LOADH, 100)]
-fn test_cuda_rand_load_sign_extend_halfword_tracegen(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = GpuChipTestBuilder::default();
-    let mut harness = create_cuda_halfword_harness(&tester);
-    for _ in 0..num_ops {
-        set_and_execute(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.dense_arena,
-            &mut rng,
-            opcode,
-            None,
-            None,
-            None,
-        );
-    }
-    transfer_load_sign_extend_records(&mut harness);
-    tester
-        .build()
-        .load_gpu_harness(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[cfg(feature = "cuda")]
-#[test_case(LOADW, 100)]
-fn test_cuda_rand_load_sign_extend_word_tracegen(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = GpuChipTestBuilder::default();
-    let mut harness = create_cuda_word_harness(&tester);
-    for _ in 0..num_ops {
-        set_and_execute(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.dense_arena,
-            &mut rng,
-            opcode,
-            None,
-            None,
-            None,
-        );
-    }
-    transfer_load_sign_extend_records(&mut harness);
-    tester
-        .build()
-        .load_gpu_harness(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
 }
