+++ extensions/riscv/circuit/src/loadstore/tests/mod.rs
@@ -24,7 +24,6 @@ use openvm_stark_backend::{
 };
 use openvm_stark_sdk::{p3_baby_bear::BabyBear, utils::create_seeded_rng};
 use rand::{rngs::StdRng, seq::IndexedRandom, Rng};
-use test_case::test_case;
 #[cfg(feature = "cuda")]
 use {
     super::{
@@ -65,6 +64,11 @@ use crate::adapters::{
 const IMM_BITS: usize = 16;
 const MAX_INS_CAPACITY: usize = 128;
 type F = BabyBear;
+
+mod byte;
+mod doubleword;
+mod halfword;
+mod word;
 
 type ByteHarness =
     TestChipHarness<F, Rv64LoadStoreByteExecutor, Rv64LoadStoreByteAir, Rv64LoadStoreByteChip<F>>;
@@ -341,309 +345,9 @@ fn memory_config_for(opcodes: &[Rv64LoadStoreOpcode]) -> MemoryConfig {
     }
     mem_config
 }
-
-#[test_case(LOADBU, 100)]
-#[test_case(STOREB, 100)]
-fn rand_loadstore_byte_test(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for(&[opcode]));
-    let (mut harness, bitwise) = create_byte_harness(&mut tester);
-    for _ in 0..num_ops {
-        set_and_execute(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.arena,
-            &mut rng,
-            opcode,
-        );
-    }
-    tester
-        .build()
-        .load(harness)
-        .load_periphery(bitwise)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[test]
-fn positive_loadwu_shift4_test() {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for(&[LOADWU]));
-    let mut harness = create_word_harness(&mut tester);
-    set_and_execute_with(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADWU,
-        Some([4, 0, 0, 0, 0, 0, 0, 0]),
-        Some(0),
-        Some(0),
-        Some(2),
-    );
-    tester
-        .build()
-        .load(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[test]
-fn positive_loadhu_shift6_test() {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for(&[LOADHU]));
-    let mut harness = create_halfword_harness(&mut tester);
-    set_and_execute_with(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADHU,
-        Some([6, 0, 0, 0, 0, 0, 0, 0]),
-        Some(0),
-        Some(0),
-        Some(2),
-    );
-    tester
-        .build()
-        .load(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[test]
-fn positive_storew_public_values_test() {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for(&[STOREW]));
-    let mut harness = create_word_harness(&mut tester);
-    set_and_execute_with(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        STOREW,
-        Some([4, 0, 0, 0, 0, 0, 0, 0]),
-        Some(0),
-        Some(0),
-        Some(PUBLIC_VALUES_AS as usize),
-    );
-    tester
-        .build()
-        .load(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[test_case(LOADHU, 100)]
-#[test_case(STOREH, 100)]
-fn rand_loadstore_halfword_test(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for(&[opcode]));
-    let mut harness = create_halfword_harness(&mut tester);
-    for _ in 0..num_ops {
-        set_and_execute(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.arena,
-            &mut rng,
-            opcode,
-        );
-    }
-    tester
-        .build()
-        .load(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[test_case(LOADWU, 100)]
-#[test_case(STOREW, 100)]
-fn rand_loadstore_word_test(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for(&[opcode]));
-    let mut harness = create_word_harness(&mut tester);
-    for _ in 0..num_ops {
-        set_and_execute(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.arena,
-            &mut rng,
-            opcode,
-        );
-    }
-    tester
-        .build()
-        .load(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[test_case(LOADD, 100)]
-#[test_case(STORED, 100)]
-fn rand_loadstore_doubleword_test(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for(&[opcode]));
-    let mut harness = create_doubleword_harness(&mut tester);
-    for _ in 0..num_ops {
-        set_and_execute(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.arena,
-            &mut rng,
-            opcode,
-        );
-    }
-    tester
-        .build()
-        .load(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
 fn b(bytes: [u8; 8]) -> [u16; BLOCK_FE_WIDTH] {
     rv64_bytes_to_u16_block(bytes)
 }
-
-#[test]
-fn run_loadd_stored_sanity_test() {
-    let read_data = b([138, 45, 202, 76, 131, 74, 186, 29]);
-    let prev_data = b([159, 213, 89, 34, 142, 67, 210, 88]);
-    assert_eq!(run_write_data(LOADD, read_data, prev_data, 0), read_data);
-    assert_eq!(run_write_data(STORED, read_data, prev_data, 0), read_data);
-}
-
-#[test]
-fn run_loadwu_storew_sanity_test() {
-    let read_data = b([138, 45, 202, 76, 131, 74, 186, 29]);
-    let prev_data = b([159, 213, 89, 34, 142, 67, 210, 88]);
-    assert_eq!(
-        run_write_data(LOADWU, read_data, prev_data, 0),
-        b([138, 45, 202, 76, 0, 0, 0, 0])
-    );
-    assert_eq!(
-        run_write_data(LOADWU, read_data, prev_data, 4),
-        b([131, 74, 186, 29, 0, 0, 0, 0])
-    );
-    assert_eq!(
-        run_write_data(STOREW, read_data, prev_data, 0),
-        b([138, 45, 202, 76, 142, 67, 210, 88])
-    );
-    assert_eq!(
-        run_write_data(STOREW, read_data, prev_data, 4),
-        b([159, 213, 89, 34, 138, 45, 202, 76])
-    );
-}
-
-#[test]
-fn run_storeh_sanity_test() {
-    let read_data = b([250, 123, 67, 198, 175, 33, 198, 250]);
-    let prev_data = b([144, 56, 175, 92, 90, 121, 64, 205]);
-    assert_eq!(
-        run_write_data(STOREH, read_data, prev_data, 0),
-        b([250, 123, 175, 92, 90, 121, 64, 205])
-    );
-    assert_eq!(
-        run_write_data(STOREH, read_data, prev_data, 2),
-        b([144, 56, 250, 123, 90, 121, 64, 205])
-    );
-    assert_eq!(
-        run_write_data(STOREH, read_data, prev_data, 4),
-        b([144, 56, 175, 92, 250, 123, 64, 205])
-    );
-    assert_eq!(
-        run_write_data(STOREH, read_data, prev_data, 6),
-        b([144, 56, 175, 92, 90, 121, 250, 123])
-    );
-}
-
-#[test]
-fn run_storeb_sanity_test() {
-    let read_data = b([221, 104, 58, 147, 175, 33, 198, 250]);
-    let prev_data = b([199, 83, 243, 12, 90, 121, 64, 205]);
-    assert_eq!(
-        run_write_data(STOREB, read_data, prev_data, 0),
-        b([221, 83, 243, 12, 90, 121, 64, 205])
-    );
-    assert_eq!(
-        run_write_data(STOREB, read_data, prev_data, 1),
-        b([199, 221, 243, 12, 90, 121, 64, 205])
-    );
-    assert_eq!(
-        run_write_data(STOREB, read_data, prev_data, 2),
-        b([199, 83, 221, 12, 90, 121, 64, 205])
-    );
-    assert_eq!(
-        run_write_data(STOREB, read_data, prev_data, 3),
-        b([199, 83, 243, 221, 90, 121, 64, 205])
-    );
-    assert_eq!(
-        run_write_data(STOREB, read_data, prev_data, 4),
-        b([199, 83, 243, 12, 221, 121, 64, 205])
-    );
-    assert_eq!(
-        run_write_data(STOREB, read_data, prev_data, 5),
-        b([199, 83, 243, 12, 90, 221, 64, 205])
-    );
-    assert_eq!(
-        run_write_data(STOREB, read_data, prev_data, 6),
-        b([199, 83, 243, 12, 90, 121, 221, 205])
-    );
-    assert_eq!(
-        run_write_data(STOREB, read_data, prev_data, 7),
-        b([199, 83, 243, 12, 90, 121, 64, 221])
-    );
-}
-
-#[test]
-fn run_loadhu_sanity_test() {
-    let read_data = b([175, 33, 198, 250, 131, 74, 186, 29]);
-    let prev_data = b([90, 121, 64, 205, 142, 67, 210, 88]);
-    assert_eq!(
-        run_write_data(LOADHU, read_data, prev_data, 0),
-        b([175, 33, 0, 0, 0, 0, 0, 0])
-    );
-    assert_eq!(
-        run_write_data(LOADHU, read_data, prev_data, 2),
-        b([198, 250, 0, 0, 0, 0, 0, 0])
-    );
-    assert_eq!(
-        run_write_data(LOADHU, read_data, prev_data, 4),
-        b([131, 74, 0, 0, 0, 0, 0, 0])
-    );
-    assert_eq!(
-        run_write_data(LOADHU, read_data, prev_data, 6),
-        b([186, 29, 0, 0, 0, 0, 0, 0])
-    );
-}
-
-#[test]
-fn run_loadbu_sanity_test() {
-    let read_data = b([131, 74, 186, 29, 138, 45, 202, 76]);
-    let prev_data = b([142, 67, 210, 88, 159, 213, 89, 34]);
-    for (shift, expected) in [
-        (0, [131, 0, 0, 0, 0, 0, 0, 0]),
-        (1, [74, 0, 0, 0, 0, 0, 0, 0]),
-        (2, [186, 0, 0, 0, 0, 0, 0, 0]),
-        (3, [29, 0, 0, 0, 0, 0, 0, 0]),
-        (4, [138, 0, 0, 0, 0, 0, 0, 0]),
-        (5, [45, 0, 0, 0, 0, 0, 0, 0]),
-        (6, [202, 0, 0, 0, 0, 0, 0, 0]),
-        (7, [76, 0, 0, 0, 0, 0, 0, 0]),
-    ] {
-        assert_eq!(
-            run_write_data(LOADBU, read_data, prev_data, shift),
-            b(expected)
-        );
-    }
-}
-
 #[test]
 fn load_sign_extend_sanity_tests() {
     let read_data = b([34, 159, 237, 151, 100, 200, 50, 25]);
@@ -694,35 +398,6 @@ fn load_sign_extend_sanity_tests() {
         b([0xAA, 0xBB, 0xCC, 0x7D, 0, 0, 0, 0])
     );
 }
-
-#[test]
-#[should_panic]
-fn solve_loadw_rejects_shift_2() {
-    run_write_data(LOADW, b([1, 2, 3, 4, 5, 6, 7, 8]), [0; BLOCK_FE_WIDTH], 2);
-}
-
-#[test]
-#[should_panic]
-fn solve_loadw_rejects_shift_6() {
-    run_write_data(LOADW, b([1, 2, 3, 4, 5, 6, 7, 8]), [0; BLOCK_FE_WIDTH], 6);
-}
-
-#[test]
-fn accepted_shift_sets() {
-    let read_data = b([0x10, 0x20, 0x30, 0x40, 0x50, 0x60, 0x70, 0x80]);
-    for shift in 0..8 {
-        let _ = run_write_data(LOADB, read_data, [0; BLOCK_FE_WIDTH], shift);
-    }
-    for shift in [0, 2, 4, 6] {
-        let _ = run_write_data(LOADH, read_data, [0; BLOCK_FE_WIDTH], shift);
-        let _ = run_write_data(LOADHU, read_data, [0; BLOCK_FE_WIDTH], shift);
-    }
-    for shift in [0, 4] {
-        let _ = run_write_data(LOADW, read_data, [0; BLOCK_FE_WIDTH], shift);
-        let _ = run_write_data(LOADWU, read_data, [0; BLOCK_FE_WIDTH], shift);
-    }
-}
-
 fn assert_pranked_byte_fails(
     opcode: Rv64LoadStoreOpcode,
     prank: impl Fn(&mut LoadStoreByteCoreCols<F>),
@@ -865,191 +540,12 @@ fn negative_split_opcode_role_tests() {
 //
 //  Ensure GPU tracegen is equivalent to CPU tracegen.
 // ////////////////////////////////////////////////////////////////////////////////////
-
-#[cfg(feature = "cuda")]
-type GpuByteHarness = GpuTestChipHarness<
-    F,
-    Rv64LoadStoreByteExecutor,
-    Rv64LoadStoreByteAir,
-    Rv64LoadStoreByteChipGpu,
-    Rv64LoadStoreByteChip<F>,
->;
-
-#[cfg(feature = "cuda")]
-type GpuHalfwordHarness = GpuTestChipHarness<
-    F,
-    Rv64LoadStoreHalfwordExecutor,
-    Rv64LoadStoreHalfwordAir,
-    Rv64LoadStoreHalfwordChipGpu,
-    Rv64LoadStoreHalfwordChip<F>,
->;
-
-#[cfg(feature = "cuda")]
-type GpuWordHarness = GpuTestChipHarness<
-    F,
-    Rv64LoadStoreWordExecutor,
-    Rv64LoadStoreWordAir,
-    Rv64LoadStoreWordChipGpu,
-    Rv64LoadStoreWordChip<F>,
->;
-
-#[cfg(feature = "cuda")]
-type GpuDoublewordHarness = GpuTestChipHarness<
-    F,
-    Rv64LoadStoreDoublewordExecutor,
-    Rv64LoadStoreDoublewordAir,
-    Rv64LoadStoreDoublewordChipGpu,
-    Rv64LoadStoreDoublewordChip<F>,
->;
-
 #[cfg(feature = "cuda")]
 fn dummy_range_checker() -> Arc<VariableRangeCheckerChip> {
     Arc::new(VariableRangeCheckerChip::new(
         default_var_range_checker_bus(),
     ))
 }
-
-#[cfg(feature = "cuda")]
-fn create_cuda_byte_harness(tester: &GpuChipTestBuilder) -> GpuByteHarness {
-    let range_checker = dummy_range_checker();
-    let bitwise_chip = Arc::new(BitwiseOperationLookupChip::<RV64_BYTE_BITS>::new(
-        default_bitwise_lookup_bus(),
-    ));
-    let air = Rv64LoadStoreByteAir::new(
-        Rv64LoadStoreAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadStoreByteCoreAir::new(
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            bitwise_chip.bus(),
-            range_checker.bus(),
-        ),
-    );
-    let executor = Rv64LoadStoreByteExecutor::new(
-        Rv64LoadStoreAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let cpu_chip = Rv64LoadStoreByteChip::<F>::new(
-        LoadStoreByteFiller::new(
-            Rv64LoadStoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            bitwise_chip,
-            range_checker,
-        ),
-        tester.dummy_memory_helper(),
-    );
-    let gpu_chip = Rv64LoadStoreByteChipGpu::new(
-        tester.range_checker(),
-        tester.bitwise_op_lookup(),
-        tester.address_bits(),
-        tester.timestamp_max_bits(),
-    );
-
-    GpuTestChipHarness::with_capacity(executor, air, gpu_chip, cpu_chip, MAX_INS_CAPACITY)
-}
-
-#[cfg(feature = "cuda")]
-fn create_cuda_halfword_harness(tester: &GpuChipTestBuilder) -> GpuHalfwordHarness {
-    let range_checker = dummy_range_checker();
-    let air = Rv64LoadStoreHalfwordAir::new(
-        Rv64LoadStoreAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadStoreHalfwordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET, range_checker.bus()),
-    );
-    let executor = Rv64LoadStoreHalfwordExecutor::new(
-        Rv64LoadStoreAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let cpu_chip = Rv64LoadStoreHalfwordChip::<F>::new(
-        LoadStoreHalfwordFiller::new(
-            Rv64LoadStoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.dummy_memory_helper(),
-    );
-    let gpu_chip = Rv64LoadStoreHalfwordChipGpu::new(
-        tester.range_checker(),
-        tester.address_bits(),
-        tester.timestamp_max_bits(),
-    );
-
-    GpuTestChipHarness::with_capacity(executor, air, gpu_chip, cpu_chip, MAX_INS_CAPACITY)
-}
-
-#[cfg(feature = "cuda")]
-fn create_cuda_word_harness(tester: &GpuChipTestBuilder) -> GpuWordHarness {
-    let range_checker = dummy_range_checker();
-    let air = Rv64LoadStoreWordAir::new(
-        Rv64LoadStoreAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadStoreWordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET, range_checker.bus()),
-    );
-    let executor = Rv64LoadStoreWordExecutor::new(
-        Rv64LoadStoreAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let cpu_chip = Rv64LoadStoreWordChip::<F>::new(
-        LoadStoreWordFiller::new(
-            Rv64LoadStoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.dummy_memory_helper(),
-    );
-    let gpu_chip = Rv64LoadStoreWordChipGpu::new(
-        tester.range_checker(),
-        tester.address_bits(),
-        tester.timestamp_max_bits(),
-    );
-
-    GpuTestChipHarness::with_capacity(executor, air, gpu_chip, cpu_chip, MAX_INS_CAPACITY)
-}
-
-#[cfg(feature = "cuda")]
-fn create_cuda_doubleword_harness(tester: &GpuChipTestBuilder) -> GpuDoublewordHarness {
-    let range_checker = dummy_range_checker();
-    let air = Rv64LoadStoreDoublewordAir::new(
-        Rv64LoadStoreAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadStoreDoublewordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET, range_checker.bus()),
-    );
-    let executor = Rv64LoadStoreDoublewordExecutor::new(
-        Rv64LoadStoreAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let cpu_chip = Rv64LoadStoreDoublewordChip::<F>::new(
-        LoadStoreDoublewordFiller::new(
-            Rv64LoadStoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.dummy_memory_helper(),
-    );
-    let gpu_chip = Rv64LoadStoreDoublewordChipGpu::new(
-        tester.range_checker(),
-        tester.address_bits(),
-        tester.timestamp_max_bits(),
-    );
-
-    GpuTestChipHarness::with_capacity(executor, air, gpu_chip, cpu_chip, MAX_INS_CAPACITY)
-}
-
 #[cfg(feature = "cuda")]
 fn transfer_loadstore_records<G, C, A, E>(harness: &mut GpuTestChipHarness<F, E, A, G, C>) {
     type Record<'a> = (&'a mut Rv64LoadStoreAdapterRecord, &'a mut LoadStoreRecord);
@@ -1060,121 +556,4 @@ fn transfer_loadstore_records<G, C, A, E>(harness: &mut GpuTestChipHarness<F, E,
             &mut harness.matrix_arena,
             EmptyAdapterCoreLayout::<F, Rv64LoadStoreAdapterExecutor>::new(),
         );
-}
-
-#[cfg(feature = "cuda")]
-#[test_case(LOADBU, 100)]
-#[test_case(STOREB, 100)]
-fn test_cuda_rand_loadstore_byte_tracegen(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester =
-        GpuChipTestBuilder::default().with_bitwise_op_lookup(default_bitwise_lookup_bus());
-    let mut harness = create_cuda_byte_harness(&tester);
-    for _ in 0..num_ops {
-        set_and_execute_with(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.dense_arena,
-            &mut rng,
-            opcode,
-            None,
-            None,
-            None,
-            Some(2),
-        );
-    }
-    transfer_loadstore_records(&mut harness);
-    tester
-        .build()
-        .load_gpu_harness(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[cfg(feature = "cuda")]
-#[test_case(LOADHU, 100)]
-#[test_case(STOREH, 100)]
-fn test_cuda_rand_loadstore_halfword_tracegen(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = GpuChipTestBuilder::default();
-    let mut harness = create_cuda_halfword_harness(&tester);
-    for _ in 0..num_ops {
-        set_and_execute_with(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.dense_arena,
-            &mut rng,
-            opcode,
-            None,
-            None,
-            None,
-            Some(2),
-        );
-    }
-    transfer_loadstore_records(&mut harness);
-    tester
-        .build()
-        .load_gpu_harness(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[cfg(feature = "cuda")]
-#[test_case(LOADWU, 100)]
-#[test_case(STOREW, 100)]
-fn test_cuda_rand_loadstore_word_tracegen(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = GpuChipTestBuilder::default();
-    let mut harness = create_cuda_word_harness(&tester);
-    for _ in 0..num_ops {
-        set_and_execute_with(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.dense_arena,
-            &mut rng,
-            opcode,
-            None,
-            None,
-            None,
-            Some(2),
-        );
-    }
-    transfer_loadstore_records(&mut harness);
-    tester
-        .build()
-        .load_gpu_harness(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
-}
-
-#[cfg(feature = "cuda")]
-#[test_case(LOADD, 100)]
-#[test_case(STORED, 100)]
-fn test_cuda_rand_loadstore_doubleword_tracegen(opcode: Rv64LoadStoreOpcode, num_ops: usize) {
-    let mut rng = create_seeded_rng();
-    let mut tester = GpuChipTestBuilder::default();
-    let mut harness = create_cuda_doubleword_harness(&tester);
-    for _ in 0..num_ops {
-        set_and_execute_with(
-            &mut tester,
-            &mut harness.executor,
-            &mut harness.dense_arena,
-            &mut rng,
-            opcode,
-            None,
-            None,
-            None,
-            Some(2),
-        );
-    }
-    transfer_loadstore_records(&mut harness);
-    tester
-        .build()
-        .load_gpu_harness(harness)
-        .finalize()
-        .simple_test()
-        .unwrap();
 }
